+++ src/source.extension.cs
@@ -9,7 +9,7 @@ namespace CloseAllTabs
     {
         public const string Id = "55640f47-34bc-436b-8820-e7f64fbb31fc";
         public const string Name = "Clean on Close";
-        public const string Description = "Closes all open documents when the solution is closing or when Visual Studio is closing";
+        public const string Description = "Closes open documents and collapses folders automatically when a solution is being closed";
         public const string Language = "en-US";
         public const string Version = "1.0";
         public const string Author = "Mads Kristensen";